requirements.txt
# Minimal core dependencies - let pip auto-resolve transitive deps
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=1.0.0
anthropic>=0.30.0
langchain-community>=0.3.0
supabase>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
resend>=2.0.0
twilio>=9.0.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.34.0

# Note: langchain, openai, httpx, pydantic, numpy will be installed as dependencies
# pandas is required by src/retrieval/career_kb.py
# requests is required for fetching live analytics from /api/analytics endpoint
# This avoids version conflicts by letting pip resolve the dependency tree

